+++ src/stc/gen_iface.py
@@ -527,20 +527,9 @@ methodOverrideMap = {
 
     'RegisterImage' :
     (0,
-     '''void %s(int type, const wxBitmap& bmp);''',
-     '''void %s(int type, const wxBitmap& bmp) {
-        // convert bmp to a xpm in a string
-        wxMemoryOutputStream strm;
-        wxImage img = bmp.ConvertToImage();
-        if (img.HasAlpha())
-            img.ConvertAlphaToMask();
-        img.SaveFile(strm, wxBITMAP_TYPE_XPM);
-        size_t len = strm.GetSize();
-        char* buff = new char[len+1];
-        strm.CopyTo(buff, len);
-        buff[len] = 0;
-        SendMsg(%s, type, (sptr_t)buff);
-        delete [] buff;'''
+     '''void %s(int type, const char* const* xpmData);''',
+     '''void %s(int type, const char* const* xpmData) {
+        SendMsg(%s, type, (sptr_t)xpmData);'''
     ),
 
     'SetHScrollBar' : ('SetUseHorizontalScrollBar', 0, 0),
